Rename to infer_type
